requirements.txt
django>=1.4
django-debug-toolbar
git+git://github.com/Brant/django-mediabrute.git
git+git://github.com/Brant/django-noodles.git
South
git+git://github.com/Brant/project_dj.git
markdown
django-taggit
mimeparse
python-dateutil
django-tastypie